App: вынес connect в отдельный модуль
