+++ src/components/app/app.test.js
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer';
-import {App} from './app';
+import App from './app';
 
 it(`App correctly renders after relaunch`, () => {
   const tree = renderer
