+++ src/components/app/index.js
@@ -1,0 +1,3 @@
+import Component from './app.connect';
+
+export default Component;
