+++ src/index.js
@@ -1,7 +1,7 @@
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import store from './store/store';
-import App from './components/app/app';
+import App from './components/app';
 import {questions, params} from './mocks/questions';
 
 const init = (gameQuestions, gameParams) => {
